backend/docu_portal_backend/requirements.txt
Flask
Flask-Cors
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-ollama
beautifulsoup4
requests
